requirements.txt
Python==3.12.7  
Matplotlib.pyplot==3.9.2 
seaborn==0.13.2
pandas==2.2.2
scikit-learn==1.5.1  
from sklearn.datasets import load_iris==1.5.1
from sklearn.model_selection import train_test_split==1.5.1
from sklearn.linear_model import LogisticRegression==1.5.1
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix==1.5.1
from sklearn.model_selection import cross_val_score, cross_val_predict==1.5.1